eilisen ja tämän päivän työt. vielä viikko 6 puolelta

+++ app/helpers/application_helper.rb
@@ -3,7 +3,11 @@ module ApplicationHelper
     return if current_user.nil?
 
     edit = link_to('Edit', url_for([:edit, item]), class: "btn btn-primary")
-    del = link_to('Destroy', item, method: :delete, data: { confirm: 'Are you sure?' }, class: "btn btn-danger")
+    del = link_to('Destroy', item, method: :delete, data: { confirm: 'Are you sure?' }, class: "btn btn-danger") if current_user.admin
     raw("#{edit} #{del}")
+  end
+
+  def round(number)
+    number_with_precision(number, precision: 1)
   end
 end
